pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-cost-analyzer"
version = "0.1.0"
description = "CLI tool that sweeps Azure subscriptions and generates interactive HTML cost reports"
requires-python = ">=3.11"
dependencies = [
    "azure-identity>=1.16.0",
    "azure-mgmt-costmanagement>=4.0.0",
    "azure-mgmt-resource>=23.0.0",
    "pandas>=2.2.0",
    "plotly>=5.22.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
pdf = ["playwright>=1.44.0"]
dev = [
    "pytest>=8.0.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
]

[project.scripts]
azure-cost-analyzer = "azure_cost_analyzer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"
